modified:   driver.c 	modified:   driver.h 	modified:   fsm.c 	modified:   fsm.h 	modified:   heis.c 	modified:   timer.c 	modified:   timer.h

+++ Kode/timer.h
@@ -3,7 +3,7 @@
 
 #include <time.h>
 
-void tmr_startTimer(float duration); // starts the timer with a duration of (duration)
+void tmr_startTimer(long double duration); // starts the timer with a duration of (duration)
 int tmr_getTimerDone(void); //returns true if the timer is done and HAS NOT BEEN READ after the timer is done
 
 #endif
